[TAN-4791] Move patch in custom_maps out of admin namespace
